せっかく5.3以上にしたので、クラス名にプリフィックスを入れるのやめてnamespaceを使うように変更
