+++ lib/Omise/OmisePHP/Account.php
@@ -4,14 +4,14 @@ namespace Omise\OmisePHP;
 
 use Omise\OmisePHP\Res\OmiseApiResourceSingleton;
 
-class OmiseBalance extends OmiseApiResourceSingleton {
-  const ENDPOINT = 'balance';
+class Account extends OmiseApiResourceSingleton {
+  const ENDPOINT = 'account';
 
   /**
-   * Retrieves a current balance in the account.
+   * Retrieves an account.
    * @param string $publickey
    * @param string $secretkey
-   * @return OmiseBalance
+   * @return OmiseAccount
    */
   public static function retrieve($publickey = null, $secretkey = null) {
     return parent::retrieve(get_class(), self::getUrl(), $publickey, $secretkey);
@@ -26,6 +26,7 @@ class OmiseBalance extends OmiseApiResourceSingleton {
   }
 
   /**
+   *
    * @param string $cardID
    * @return string
    */
