+++ lib/Omise/OmisePHP/Balance.php
@@ -4,14 +4,14 @@ namespace Omise\OmisePHP;
 
 use Omise\OmisePHP\Res\OmiseApiResourceSingleton;
 
-class OmiseAccount extends OmiseApiResourceSingleton {
-  const ENDPOINT = 'account';
+class Balance extends OmiseApiResourceSingleton {
+  const ENDPOINT = 'balance';
 
   /**
-   * Retrieves an account.
+   * Retrieves a current balance in the account.
    * @param string $publickey
    * @param string $secretkey
-   * @return OmiseAccount
+   * @return OmiseBalance
    */
   public static function retrieve($publickey = null, $secretkey = null) {
     return parent::retrieve(get_class(), self::getUrl(), $publickey, $secretkey);
@@ -26,7 +26,6 @@ class OmiseAccount extends OmiseApiResourceSingleton {
   }
 
   /**
-   *
    * @param string $cardID
    * @return string
    */
