+++ tests/Omise/OmisePHP/BalanceTest.php
@@ -2,7 +2,7 @@
 
 namespace Omise\OmisePHP\Tests;
 
-use Omise\OmisePHP\OmiseBalance;
+use Omise\OmisePHP;
 
 define('OMISE_PUBLIC_KEY', 'pkey');
 define('OMISE_SECRET_KEY', 'skey');
@@ -21,7 +21,7 @@ class OmiseBalanceTest extends \PHPUnit_Framework_TestCase {
    * reloadに成功し、objectの値がbalanceであれば正しいとみなす
    */
   public function testReload() {
-    $balance = OmiseBalance::retrieve();
+    $balance = OmisePHP\Balance::retrieve();
     $balance->reload();
 
     $this->assertArrayHasKey('object', $balance);
@@ -33,7 +33,7 @@ class OmiseBalanceTest extends \PHPUnit_Framework_TestCase {
    * retrieveに成功し、objectの値がbalanceであれば正しいとみなす
    */
   public function testRetrieve() {
-    $balance = OmiseBalance::retrieve();
+    $balance = OmisePHP\Balance::retrieve();
     
     $this->assertArrayHasKey('object', $balance);
     $this->assertEquals('balance', $balance['object']);
@@ -44,8 +44,8 @@ class OmiseBalanceTest extends \PHPUnit_Framework_TestCase {
    * retrieveを2度実行した結果のインスタンスが同一であれば正しいとみなす
    */
   public function testSameInstance() {
-    $balance1 = OmiseBalance::retrieve();
-    $balance2 = OmiseBalance::retrieve();
+    $balance1 = OmisePHP\Balance::retrieve();
+    $balance2 = OmisePHP\Balance::retrieve();
     
     $this->assertTrue($balance1 === $balance2);
   }
